requirements.txt
Pillow==10.4.0
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1